+++ app/src/organisms/Devices/HistoricalProtocolRun.tsx
@@ -17,13 +17,11 @@ import { StyledText } from '../../atoms/text'
 import { getStoredProtocols } from '../../redux/protocol-storage'
 import { formatInterval } from '../RunTimeControl/utils'
 import { formatTimestamp } from './utils'
-import { useRunTimestamps } from '../RunTimeControl/hooks'
+import { EMPTY_TIMESTAMP } from './constants'
 import { HistoricalProtocolRunOverflowMenu as OverflowMenu } from './HistoricalProtocolRunOverflowMenu'
 import { HistoricalProtocolRunOffsetDrawer as OffsetDrawer } from './HistoricalProtocolRunOffsetDrawer'
 import type { RunData } from '@opentrons/api-client'
 import type { State } from '../../redux/types'
-
-const EMPTY_TIMESTAMP = '--:--:--'
 
 const CLICK_STYLE = css`
   cursor: pointer;
@@ -46,7 +44,6 @@ export function HistoricalProtocolRun(
   const { run, protocolName, robotIsBusy, robotName, protocolKey, key } = props
   const history = useHistory()
   const [offsetDrawerOpen, setOffsetDrawerOpen] = React.useState(false)
-  const { startedAt, completedAt } = useRunTimestamps(run.id)
   const storedProtocols = useSelector((state: State) =>
     getStoredProtocols(state)
   )
@@ -54,10 +51,10 @@ export function HistoricalProtocolRun(
   const runDisplayName = formatTimestamp(run.createdAt)
   let duration = EMPTY_TIMESTAMP
   if (runStatus !== 'idle') {
-    if (completedAt != null && startedAt != null) {
-      duration = formatInterval(startedAt, completedAt)
-    } else if (startedAt != null) {
-      duration = formatInterval(startedAt, new Date().toString())
+    if (run.completedAt != null && run.startedAt != null) {
+      duration = formatInterval(run.startedAt, run.completedAt)
+    } else if (run.startedAt != null) {
+      duration = formatInterval(run.startedAt, new Date().toString())
     }
   }
   const protocolKeyInStoredKeys = storedProtocols.find(